Fix for task id wrap round
git-svn-id: https://mydns-ng.svn.sourceforge.net/svnroot/mydns-ng@164 fff143f6-a9c2-4b89-8f93-78cb2c0259b0

+++ mydns/branches/howard/pre-1.2.9/src/lib/taskobj.c
@@ -38,7 +38,7 @@
 #define TASKVEC_ISSET(TI, TV)	FD_ISSET((TI), (fd_set*)(TV))
 
 static uint8_t		*taskvec = NULL;
-static uint16_t		internal_id = 0;
+static uint32_t		internal_id = 0;
 static int32_t		active_tasks = 0;
 
 uint32_t 	answer_then_quit = 0;		/* Answer this many queries then quit */
